Renamed file for User Settings page and fixed routing
