+++ client/src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import './App.css';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import { LandingPage } from './components/LandingPage/LandingPage';
-import { UserSettingsPage } from './components/UserSettingsPage/UserSettingsPage';
+import { UserSettingsPage } from './components/UserSettingsPage';
 
 function App() {
   return (
